refactor: extract shared markdown utilities and deduplicate copy button
