+++ packages/frontend/src/components/ui/markdown-preview-modal.tsx
@@ -1,5 +1,4 @@
-import { type MouseEvent, type ReactElement, useCallback } from "react";
-import { CopyIconButton } from "@/components/ui/copy-icon-button";
+import type { ReactElement } from "react";
 import {
   Dialog,
   DialogBody,
@@ -8,9 +7,9 @@ import {
   DialogHeader,
   DialogTitle,
 } from "@/components/ui/dialog";
+import { DocumentCopyButton } from "@/components/ui/document-copy-button";
 import { MarkdownRenderer } from "@/components/ui/markdown-renderer";
-import { buildCopyPreview } from "@/lib/copy-preview";
-import { useCopyToClipboard } from "@/lib/use-copy-to-clipboard";
+import { hasLabeledCodeFence } from "@/lib/markdown-utils";
 
 export type MarkdownPreviewModalProps = {
   markdown: string;
@@ -18,36 +17,6 @@ export type MarkdownPreviewModalProps = {
   onOpenChange: (open: boolean) => void;
   title?: string;
 };
-
-const hasLabeledCodeFence = (markdown: string): boolean => {
-  return markdown.includes("```") && /```[a-z0-9_-]+/i.test(markdown);
-};
-
-function MarkdownPreviewModalCopyButton({ markdown }: { markdown: string }): ReactElement {
-  const { copied, copyToClipboard } = useCopyToClipboard({
-    getSuccessDescription: buildCopyPreview,
-    errorLogContext: "MarkdownPreviewModal",
-  });
-
-  const handleCopy = useCallback(
-    (event: MouseEvent<HTMLButtonElement>) => {
-      event.stopPropagation();
-      event.preventDefault();
-      void copyToClipboard(markdown);
-    },
-    [copyToClipboard, markdown],
-  );
-
-  return (
-    <CopyIconButton
-      copied={copied}
-      ariaLabel="Copy document content"
-      dataTestId="markdown-preview-modal-copy"
-      className="absolute top-2 right-2 z-10"
-      onClick={handleCopy}
-    />
-  );
-}
 
 export function MarkdownPreviewModal({
   markdown,
@@ -73,7 +42,12 @@ export function MarkdownPreviewModal({
               variant="document"
               premiumCodeBlocks={hasLabeledCodeFence(markdown)}
             />
-            <MarkdownPreviewModalCopyButton markdown={markdown} />
+            <DocumentCopyButton
+              markdown={markdown}
+              dataTestId="markdown-preview-modal-copy"
+              errorLogContext="MarkdownPreviewModal"
+              className="absolute top-2 right-2 z-10"
+            />
           </div>
         </DialogBody>
       </DialogContent>
