+++ packages/frontend/src/lib/markdown-utils.ts
@@ -1,0 +1,3 @@
+export const hasLabeledCodeFence = (markdown: string): boolean => {
+  return markdown.includes("```") && /```[a-z0-9_-]+/i.test(markdown);
+};
